+++ mycodo/functions/average_past_single.py
@@ -61,7 +61,7 @@ measurements_dict = {
 }
 
 FUNCTION_INFORMATION = {
-    'function_name_unique': 'average_past_single',
+    'function_name_unique': 'AVG_PAST_SINGLE',
     'function_name': 'Average (Past, Single)',
     'measurements_dict': measurements_dict,
     'enable_channel_unit_select': True,